requirements.txt
numpy
pyOpenSSL
watchdog
opencv-python-headless
matplotlib
openai
torchaudio

clip-interrogator==0.6.0
transformers>=4.36.0
lark-parser
imageio-ffmpeg
rembg[gpu]
omegaconf>=2.3.0
Pillow>=9.5.0
einops>=0.7.0
trimesh>=4.0.5
huggingface-hub
scikit-image
soundfile>=0.12.1
json-repair

bitsandbytes
accelerate
scenedetect[opencv-headless]

hydra-core>=1.3.2
loralib>=0.1.2
natsort>=8.4.0
# simple-lama-inpainting